src/back_end/requirements.txt
bottle          # wsgi framework
requests
jsonschema
coloredlogs
wikidata
pyyaml
